+++ BasicUI2/src/modules/highgui/src/cap_ios_abstract_camera.mm
@@ -242,7 +242,7 @@
         default:
             break;
     }
-    NSLog(@"deviceOrientationDidChange: %d", orientation);
+    NSLog(@"deviceOrientationDidChange: %ld", (long)orientation);
 
     [self updateOrientation];
 }
